Ziua 7, luni 26.06

+++ cleanup.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+#default goleste cahe-ul pentru un utlizator, fisierul comun pentru fisiere temp si fisierele cu extensia .log din var/log
+
 if [[ ! $1 ]]
 then
     echo "NU ai trimis niciun argument!"
@@ -14,10 +16,17 @@ file=`ls /home |egrep -ow "temp_files"`
 if [[ ! $file ]]
 then
     sudo mkdir /home/temp_files
+    chmod 777 /home/temp_files
     echo "A fost creat un director comun pentru utilizatori, pentru fisiere temporare!"
 else
-   sudo rm -rvf /home/temp_files/*
-   echo "/home/temp_files a fost golit!"
+    cond=`ls /home/temp_files`
+    if [[ $cond ]]
+    then
+        sudo rm -rvf /home/temp_files/*
+        echo "/home/temp_files a fost golit!"
+    else
+        echo "temp_files este deja gol!"
+    fi
 fi
